Extract mobile pending sync hook
